Fix bugs on cost calculation

+++ Com.Bateeq.Service.Merchandiser.Lib/Services/CostCalculationRetailService.cs
@@ -333,7 +333,7 @@ namespace Com.Bateeq.Service.Merchandiser.Lib.Services
             model.SH_Cutting = viewModel.SH_Cutting != null ? (double)viewModel.SH_Cutting : 0;
             model.SH_Sewing = viewModel.SH_Sewing != null ? (double)viewModel.SH_Sewing : 0;
             model.SH_Finishing = viewModel.SH_Finishing != null ? (double)viewModel.SH_Finishing : 0;
-            model.Quantity = viewModel.Quantity != null ? (double)viewModel.Quantity : 0;
+            model.Quantity = viewModel.Quantity != null ? (int)viewModel.Quantity : 0;
             model.EfficiencyId = viewModel.Efficiency.Id != null ? (int)viewModel.Efficiency.Id : 0;
             model.EfficiencyValue = viewModel.Efficiency.Value != null ? (double)viewModel.Efficiency.Value : 0;
             model.Risk = viewModel.Risk != null ? (double)viewModel.Risk / 100 : 0;
@@ -387,7 +387,7 @@ namespace Com.Bateeq.Service.Merchandiser.Lib.Services
                 costCalculationRetail_Material.CategoryName = costCalculationRetail_MaterialVM.Category.Name + " - " + costCalculationRetail_MaterialVM.Category.SubCategory;
                 costCalculationRetail_Material.MaterialId = costCalculationRetail_MaterialVM.Material.Id != null ? (int)costCalculationRetail_MaterialVM.Material.Id : 0;
                 costCalculationRetail_Material.MaterialName = costCalculationRetail_MaterialVM.Material.Name;
-                costCalculationRetail_Material.Quantity = costCalculationRetail_MaterialVM.Quantity != null ? (int)costCalculationRetail_MaterialVM.Quantity : 0;
+                costCalculationRetail_Material.Quantity = costCalculationRetail_MaterialVM.Quantity != null ? (double)costCalculationRetail_MaterialVM.Quantity : 0;
                 costCalculationRetail_Material.UOMQuantityId = costCalculationRetail_MaterialVM.UOMQuantity.Id != null ? (int)costCalculationRetail_MaterialVM.UOMQuantity.Id : 0;
                 costCalculationRetail_Material.UOMQuantityName = costCalculationRetail_MaterialVM.UOMQuantity.Name;
                 costCalculationRetail_Material.Price = costCalculationRetail_MaterialVM.Price != null ? (double)costCalculationRetail_MaterialVM.Price : 0;
